requirements.txt
@"
# Core Testing Frameworks
pytest==7.4.3
playwright==1.40.0
selenium==4.15.2

# Pytest Plugins
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-rerunfailures==12.0
pytest-cov==4.1.0
pytest-json-report==1.5.0
allure-pytest==2.13.2

# Browser Drivers
webdriver-manager==4.0.1

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
faker==20.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
pylint==3.0.3
"@ | Out-File -FilePath "requirements.txt" -Encoding UTF8